refactor(lineSegment): extract module-level helpers to utils/lineSegment
Move the slope/length/midpoint helper functions from the top of
LineSegment.js into a dedicated utils/lineSegment.js module. Export the
nine helpers the component uses and keep the two internal-only helpers
(getNumericValue, getSlopeAndSignedLength) private to the module.

No behavior change.
